+++ src/rendering/gs_rasterizer.hpp
@@ -1,14 +1,12 @@
 #pragma once
 
 #include "core/camera.hpp"
-#include "core/rasterizer.hpp"
 #include "core/splat_data.hpp"
-#include "rasterization_api.h"
+#include <tuple>
 
 namespace gs::rendering {
 
-    // Wrapper function to use fastgs backend for rendering
-    RenderOutput fast_rasterize(
+    torch::Tensor rasterize(
         Camera& viewpoint_camera,
         SplatData& gaussian_model,
         torch::Tensor& bg_color);
